+++ lib/core/routing/app_router.dart
@@ -270,5 +270,6 @@ class AppRouter {
       default:
         return null;
     }
+    return null;
   }
 }
