--- lib/core/widgets/custom_app_bar.dart
@@ -1,38 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:healthstack/core/theming/colors.dart';
-import 'package:healthstack/core/theming/styles.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-
-class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
-  final String title;
-
-  const CustomAppBar({
-    super.key,
-    required this.title,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return AppBar(
-      elevation: 0,
-      centerTitle: true,
-      backgroundColor: Colors.white,
-      title: Text(
-        title,
-        style: TextStyles.font20DarkBlueSemiBold,
-      ),
-      
-      leading: InkWell(
-        onTap: () => Navigator.pop(context),
-        child: Icon(
-          Icons.arrow_back_ios_new_outlined,
-          color: ColorsManager.mainBlue,
-          size: 20.sp,
-        ),
-      ),
-    );
-  }
-
-  @override
-  Size get preferredSize => const Size.fromHeight(kToolbarHeight);
-}
